+++ Uplift/Models/UserProfile.swift
@@ -1,0 +1,14 @@
+//
+//  UserProfile.swift
+//  Uplift
+//
+//  Created by jiwon jeong on 3/16/25.
+//  Copyright © 2025 Cornell AppDev. All rights reserved.
+//
+
+import Foundation
+
+struct UserProfile {
+    let id: String
+    let name: String
+}
