+++ Uplift/Models/WorkoutHistory.swift
@@ -1,0 +1,16 @@
+//
+//  WorkoutHistory.swift
+//  Uplift
+//
+//  Created by jiwon jeong on 3/16/25.
+//  Copyright © 2025 Cornell AppDev. All rights reserved.
+//
+
+import Foundation
+
+struct WorkoutHistory: Identifiable {
+    let id: String
+    let location: String
+    let time: String
+    let date: String
+}
